+++ changelog.md
@@ -3,7 +3,14 @@
 Data:
 Data Realizado:
 Local: Contas/Cadastro
-Alteração: 
+Alteração:
 
 * [ ] Alterar a posição entre Parcelas e Total, ficando Total e Parcelas
 * [ ] Ao selecionar a parcela, preencher os campos com a divição total com com as parcelas
+
+---
+
+* [ ] Login
+* [ ] Níveis
+* [ ] Alterar cadastro de usuario para CPF
+* [ ]